app/api-tenants/requirements.txt
fastapi
uvicorn
pydantic
azure-cosmos